Backend/requirements.txt
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
nanoid==2.0.0
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.20.0
Werkzeug==3.0.4
